chore: rename NavItem.tsx -> Menu.tsx
